+++ lab5/functions.cpp
@@ -1,6 +1,6 @@
 #include "functions.h"
 
-double f(double x, double y, double t)
+double f(double x, double y, int t)
 {
     return y * cos(t * x);
 }
